backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic[email]==2.7.1
passlib[bcrypt,argon2]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
python-dotenv==1.0.1
pyotp==2.9.0
argon2-cffi==23.1.0
aioboto3==13.0.0
openai==1.30.1
slowapi==0.1.9
websockets==12.0
httpx==0.27.0
motor==3.4.0
pymongo==4.6.3
cryptography==42.0.5
